refactor update logic
