+++ admin-portal/src/pages/aidPackage/aidPackage.tsx
@@ -5,6 +5,7 @@ import { Page } from "layout/page";
 import { AssignSuppliers } from "./assignSuppliers/assignSuppliers";
 import { ManageAidPackages } from "./manageAidPackages/manageAidPackages";
 import { fetchMedicalNeeds, NeedsInfo } from "data/medical-needs.mock.data";
+import AdminDataServices from "../../apis/services";
 
 import "./aidPackage.css";
